Add support for non-char basic_string/basic_string_view
We currently have cases where we are using `ustring` or `bstring` (or
similar views), which currently have to get converted to plain
std::string_views.

On *output* we also have trouble where our dict producer is building an
std::string but we need an ustring (and so have to reallocate another
string and copy the bytes over).

This fixes it by templatizing bt producers and consumers to allow any
size-1 char type to be used.