Add append_encoded to list/dict producers
Sometimes we have an already-encoded bt-dict and don't know all of the
keys, but don't want to fully deserialize/serialize it to not lose
unknown keys.  This allows one to append a string that represents
already encoded bt data, e.g. "d1:a3:fooe" as-is, rather than as a
string ("10:d1:a3:fooe"), making the result more correct (and smaller).

Debug builds will confirm that the data you're appending is valid
bt-encoded data, but release builds will skip this check.  Caveat
emptor!